pyproject.toml
[project]
name = "obsidian-web-mcp"
version = "0.1.0"
description = "Secure remote MCP server for Obsidian vaults -- access your notes from Claude, your phone, or any MCP client, anywhere"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.0",
    "python-frontmatter>=1.1.0",
    "pydantic>=2.0",
    "watchdog>=4.0.0",
    "fastembed>=0.8.0",
    "sqlite-vec>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
vault-mcp = "obsidian_vault_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_vault_mcp"]
